npm uninstall aframe because of vulnerability
