+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
+import Script from 'next/script'
 import './globals.css'
 
 const inter = Inter({ subsets: ['latin'] })
@@ -17,6 +18,17 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>{children}</body>
+      <Script
+          src="https://aframe.io/releases/1.5.0/aframe.min.js"
+      />
+      <Script
+        src="play-on-click.js"
+        strategy="lazyOnload"
+      />
+      <Script
+        src="hide-on-play.js"
+        strategy="lazyOnload"
+      />
     </html>
   )
 }
