+++ package.json
@@ -9,8 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/aframe": "^1.2.6",
-    "aframe": "^1.5.0",
     "aframe-react": "^4.4.0",
     "leaflet": "^1.9.4",
     "next": "^14.1.0",
